create index from length
this will create the index directly from the length of the list.

That is because the CSV file currently doesn't contain an internal index, so if I import it the result will be useless.

However, if I do this, the class generates it, and then can save it into the file, thereby fixing the input